大量代码重构、添加图片上传方法

+++ EM-Pojo/src/main/java/com/emt/composite/EventComposite.java
@@ -11,6 +11,6 @@ import java.math.BigInteger;
 public class EventComposite extends Event {
     private String categoryName;
     private BigDecimal ratingValue;
-    private BigInteger participantNum;
+    private Integer participantNum;
     private Boolean status;
 }
